requirements.txt
setuptools
pyaudio
vosk
SpeechRecognition
pocketsphinx
openai
pyttsx3
pillow
pywin32
pyautogui